Move api constants
