+++ frontend/src/admin/Admin.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Admin as ReactAdmin, Resource, ListGuesser, EditGuesser, fetchUtils } from 'react-admin';
 import jsonServerProvider from 'ra-data-json-server';
-import { API_BASE_URL } from 'utils/constants';
+import { API_BASE_URL } from 'utils/constants/apiConstants';
 import { QuestionList } from 'components/QuestionList';
 import { QuestionEdit } from 'components/QuestionEdit';
 import { CategoryCreate } from 'components/CategoryCreate';
