+++ frontend/src/utils/constants/apiConstants.js
@@ -1,4 +1,4 @@
-import config from './config.json';
+import config from 'utils/config.json';
 
 export const API_BASE_URL = config[process.env.NODE_ENV].BACKEND_URL;
